Move AdministrationPrivilege to admin package (#9717)
